clarify OpenPGP data handling in "about"
The term "technical data" bothered me, since all of the identity
information (including e-mail addresses!) are technical data too.

The "up- and downloaded on" construction was not only
awkward, it was incorrect: the policies for what can be
uploaded differ from the policies of what will be downloaded.
Separating out what will be stored and retransmitted seems better.

The last "However," statement wasn't really a "however" -- it's one of
the main reasons for the keyserver to exist! I've made it a bit
stronger.