Simplify and clarify description of technical data.
shorter sentences.

"verifying signatures and encrypting messages" is clearer than
"cryptographic operations".

There is no "but" here -- this is a list of things that are included.

not all expiration dates are necessarily included here, since the
expiration dates of the user IDs (and therefore of the certificate
itself in some contexts) are in self-sigs.

"whether it is revoked or not" is too broad of a brush -- what does
"it" refer to?  since subkeys can be independently revoked, we're
talking about a piecemeal revocation status.

Mention the creation date, since that's one of the few things that is
really always going to be transmitted alongside the primary key
materal.