recursively propagate changes
We now publish mutual certifications.  Consider Alice certifies Bob,
and Bob certifies Alice.  When Alice uploads her cert, Bob's
certifications are not published because we don't know that the
certification is mutual because we haven't seen Bob's cert yet.  When
Bob uploads his cert, we now see that the certifications are mutual,
and can immediately publish Alice's certifications on Bob's cert.

However, we also need to update Alice's cert to now include Bob's
certification.  This patch implements that by propagating changes up
to a certain depth (currently 1).  We release the database lock as not
to starve other requests.